+++ doubly_linked_list/doubly_linked_list.py
