+++ linked_list/linked_list.py
